requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-multipart==0.0.18
supabase==2.11.0
python-dotenv==1.0.1
itsdangerous==2.2.0
anthropic==0.43.0
